+++ src/components/spatial/SpatialSubscriber.js
@@ -35,6 +35,7 @@ export default function SpatialSubscriber({
   moleculeRadius,
   view,
   cellRadius,
+  observationsLabelOverride,
   theme,
   uuid = null,
 }) {
@@ -187,11 +188,12 @@ export default function SpatialSubscriber({
     layerName => PubSub.publish(CLEAR_PLEASE_WAIT, layerName),
     [],
   );
+  const observationsLabel = observationsLabelOverride || 'cells';
   return (
     <TitleInfo
       title="Spatial"
       info={
-        `${cellsCount} cells, ${moleculesCount} molecules at ${shortNumber(locationsCount)} locations`
+        `${cellsCount} ${observationsLabel}, ${moleculesCount} molecules at ${shortNumber(locationsCount)} locations`
       }
       isSpatial
       urls={urls}